move template agent files to a seperate packate
