+++ pkg/manager/templateagent/registration.go
@@ -1,0 +1,434 @@
+package templateagent
+
+import (
+	"context"
+	"crypto/x509"
+	"encoding/pem"
+	"fmt"
+	"strings"
+	"time"
+
+	openshiftcrypto "github.com/openshift/library-go/pkg/crypto"
+	"github.com/pkg/errors"
+	certificatesv1 "k8s.io/api/certificates/v1"
+	rbacv1 "k8s.io/api/rbac/v1"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
+	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
+	"k8s.io/client-go/kubernetes"
+	rbacv1lister "k8s.io/client-go/listers/rbac/v1"
+	"k8s.io/klog/v2"
+	addonapiv1alpha1 "open-cluster-management.io/api/addon/v1alpha1"
+	addonlisterv1alpha1 "open-cluster-management.io/api/client/addon/listers/addon/v1alpha1"
+	clusterv1 "open-cluster-management.io/api/cluster/v1"
+
+	"open-cluster-management.io/addon-framework/pkg/agent"
+	"open-cluster-management.io/addon-framework/pkg/utils"
+)
+
+const (
+	TLSCACert = "ca.crt"
+	TLSCAKey  = "ca.key"
+)
+
+// AddonTemplateGetterFunc returns the desired addon template for the addon, if addon is nil, the clusterName and
+// addonName should not be empty
+type AddonTemplateGetterFunc func(addon *addonapiv1alpha1.ManagedClusterAddOn, clusterName, addonName string) (*addonapiv1alpha1.AddOnTemplate, error)
+
+// DefaultDesiredAddonTemplateGetter returns a function that returns the desired addon template
+// if addon is nil, it will get the addon from cache, then get the desired template by addon status
+func DefaultDesiredAddonTemplateGetter(
+	mcaLister addonlisterv1alpha1.ManagedClusterAddOnLister,
+	atLister addonlisterv1alpha1.AddOnTemplateLister) AddonTemplateGetterFunc {
+	return func(addon *addonapiv1alpha1.ManagedClusterAddOn, clusterName, addonName string) (*addonapiv1alpha1.AddOnTemplate, error) {
+
+		if addon == nil {
+			var err error
+			addon, err = mcaLister.ManagedClusterAddOns(clusterName).Get(addonName)
+			if err != nil {
+				return nil, err
+			}
+		}
+
+		return GetDesiredAddOnTemplate(atLister, addon)
+	}
+}
+
+func TemplateCSRConfigurationsFunc(
+	addonName, agentName string,
+	templateGetter AddonTemplateGetterFunc,
+) func(cluster *clusterv1.ManagedCluster) []addonapiv1alpha1.RegistrationConfig {
+
+	return func(cluster *clusterv1.ManagedCluster) []addonapiv1alpha1.RegistrationConfig {
+		template, err := templateGetter(nil, cluster.Name, addonName)
+		if err != nil {
+			utilruntime.HandleError(fmt.Errorf("failed to get addon %s template: %v", addonName, err))
+			return nil
+		}
+		if template == nil {
+			return nil
+		}
+
+		contain := func(rcs []addonapiv1alpha1.RegistrationConfig, signerName string) bool {
+			for _, rc := range rcs {
+				if rc.SignerName == signerName {
+					return true
+				}
+			}
+			return false
+		}
+
+		registrationConfigs := make([]addonapiv1alpha1.RegistrationConfig, 0)
+		for _, registration := range template.Spec.Registration {
+			switch registration.Type {
+			case addonapiv1alpha1.RegistrationTypeKubeClient:
+				if !contain(registrationConfigs, certificatesv1.KubeAPIServerClientSignerName) {
+					configs := agent.KubeClientSignerConfigurations(addonName, agentName)(cluster)
+					registrationConfigs = append(registrationConfigs, configs...)
+				}
+
+			case addonapiv1alpha1.RegistrationTypeCustomSigner:
+				if registration.CustomSigner == nil {
+					continue
+				}
+				if !contain(registrationConfigs, registration.CustomSigner.SignerName) {
+					configs := CustomSignerConfigurations(
+						addonName, agentName, registration.CustomSigner)(cluster)
+					registrationConfigs = append(registrationConfigs, configs...)
+				}
+
+			default:
+				utilruntime.HandleError(fmt.Errorf("unsupported registration type %s", registration.Type))
+			}
+
+		}
+
+		return registrationConfigs
+	}
+}
+
+// CustomSignerConfigurations returns a func that can generate RegistrationConfig
+// for CustomSigner type registration addon
+func CustomSignerConfigurations(addonName, agentName string,
+	customSignerConfig *addonapiv1alpha1.CustomSignerRegistrationConfig,
+) func(cluster *clusterv1.ManagedCluster) []addonapiv1alpha1.RegistrationConfig {
+	return func(cluster *clusterv1.ManagedCluster) []addonapiv1alpha1.RegistrationConfig {
+		if customSignerConfig == nil {
+			utilruntime.HandleError(fmt.Errorf("custome signer is nil"))
+		}
+		config := addonapiv1alpha1.RegistrationConfig{
+			SignerName: customSignerConfig.SignerName,
+			// TODO: confirm the subject
+			Subject: addonapiv1alpha1.Subject{
+				User:   agent.DefaultUser(cluster.Name, addonName, agentName),
+				Groups: agent.DefaultGroups(cluster.Name, addonName),
+			},
+		}
+		if customSignerConfig.Subject != nil {
+			config.Subject = *customSignerConfig.Subject
+		}
+
+		return []addonapiv1alpha1.RegistrationConfig{config}
+	}
+}
+
+func TemplateCSRApproveCheckFunc(
+	addonName, agentName string,
+	templateGetter AddonTemplateGetterFunc,
+) agent.CSRApproveFunc {
+
+	return func(cluster *clusterv1.ManagedCluster, addon *addonapiv1alpha1.ManagedClusterAddOn,
+		csr *certificatesv1.CertificateSigningRequest) bool {
+
+		template, err := templateGetter(addon, cluster.Name, addonName)
+		if err != nil {
+			utilruntime.HandleError(fmt.Errorf("failed to get addon %s template: %v", addonName, err))
+			return false
+		}
+		if template == nil {
+			return false
+		}
+
+		for _, registration := range template.Spec.Registration {
+			switch registration.Type {
+			case addonapiv1alpha1.RegistrationTypeKubeClient:
+
+				if csr.Spec.SignerName == certificatesv1.KubeAPIServerClientSignerName {
+					return KubeClientCSRApprover(agentName)(cluster, addon, csr)
+				}
+
+			case addonapiv1alpha1.RegistrationTypeCustomSigner:
+				if registration.CustomSigner == nil {
+					continue
+				}
+				if csr.Spec.SignerName == registration.CustomSigner.SignerName {
+					return CustomerSignerCSRApprover(addonName)(cluster, addon, csr)
+				}
+
+			default:
+				utilruntime.HandleError(fmt.Errorf("unsupported registration type %s", registration.Type))
+			}
+
+		}
+
+		return false
+	}
+}
+
+// KubeClientCSRApprover approve the csr when addon agent uses default group, default user and
+// "kubernetes.io/kube-apiserver-client" signer to sign csr.
+func KubeClientCSRApprover(agentName string) agent.CSRApproveFunc {
+	return func(
+		cluster *clusterv1.ManagedCluster,
+		addon *addonapiv1alpha1.ManagedClusterAddOn,
+		csr *certificatesv1.CertificateSigningRequest) bool {
+		if csr.Spec.SignerName != certificatesv1.KubeAPIServerClientSignerName {
+			return false
+		}
+		return utils.DefaultCSRApprover(agentName)(cluster, addon, csr)
+	}
+}
+
+// CustomerSignerCSRApprover approve the csr when addon agent uses custom signer to sign csr.
+func CustomerSignerCSRApprover(agentName string) agent.CSRApproveFunc {
+	return func(
+		cluster *clusterv1.ManagedCluster,
+		addon *addonapiv1alpha1.ManagedClusterAddOn,
+		csr *certificatesv1.CertificateSigningRequest) bool {
+
+		klog.Infof("Customer signer CSR is approved. cluster: %s, addon %s, requester: %s",
+			cluster.Name, addon.Name, csr.Spec.Username)
+		return true
+	}
+}
+
+func TemplateCSRSignFunc(
+	addonName, agentName string,
+	templateGetter AddonTemplateGetterFunc,
+	hubKubeClient kubernetes.Interface,
+) agent.CSRSignerFunc {
+
+	return func(csr *certificatesv1.CertificateSigningRequest) []byte {
+		// TODO: consider to change the agent.CSRSignerFun to accept parameter addon
+		getClusterName := func(userName string) string {
+			// the common name of csr is in format of "system:open-cluster-management:{clusterName}:{id}"
+			// get the cluster name from the common name
+			trimedCommonName := strings.TrimPrefix(userName, "system:open-cluster-management:")
+			return strings.Split(trimedCommonName, ":")[0]
+		}
+
+		clusterName := getClusterName(csr.Spec.Username)
+		template, err := templateGetter(nil, clusterName, addonName)
+		if err != nil {
+			utilruntime.HandleError(fmt.Errorf("failed to get addon %s template: %v", addonName, err))
+			return nil
+		}
+		if template == nil {
+			return nil
+		}
+
+		for _, registration := range template.Spec.Registration {
+			switch registration.Type {
+			case addonapiv1alpha1.RegistrationTypeKubeClient:
+				continue
+
+			case addonapiv1alpha1.RegistrationTypeCustomSigner:
+				if registration.CustomSigner == nil {
+					continue
+				}
+				if csr.Spec.SignerName == registration.CustomSigner.SignerName {
+					return CustomSignerWithExpiry(hubKubeClient, registration.CustomSigner, 24*time.Hour)(csr)
+				}
+
+			default:
+				utilruntime.HandleError(fmt.Errorf("unsupported registration type %s", registration.Type))
+			}
+
+		}
+
+		return nil
+	}
+}
+
+func CustomSignerWithExpiry(
+	kubeclient kubernetes.Interface,
+	customSignerConfig *addonapiv1alpha1.CustomSignerRegistrationConfig,
+	duration time.Duration) agent.CSRSignerFunc {
+	return func(csr *certificatesv1.CertificateSigningRequest) []byte {
+		if customSignerConfig == nil {
+			utilruntime.HandleError(fmt.Errorf("custome signer is nil"))
+			return nil
+		}
+
+		if csr.Spec.SignerName != customSignerConfig.SignerName {
+			return nil
+		}
+		caSecret, err := kubeclient.CoreV1().Secrets(customSignerConfig.SigningCA.Namespace).Get(
+			context.TODO(), customSignerConfig.SigningCA.Name, metav1.GetOptions{})
+		if err != nil {
+			utilruntime.HandleError(fmt.Errorf("get custome signer ca %s/%s failed, %v",
+				customSignerConfig.SigningCA.Namespace, customSignerConfig.SigningCA.Name, err))
+			return nil
+		}
+
+		caData, caKey, err := extractCAdata(caSecret.Data[TLSCACert], caSecret.Data[TLSCAKey])
+		if customSignerConfig == nil {
+			utilruntime.HandleError(fmt.Errorf("get ca %s/%s data failed, %v",
+				customSignerConfig.SigningCA.Namespace, customSignerConfig.SigningCA.Name, err))
+			return nil
+		}
+		return utils.DefaultSignerWithExpiry(caKey, caData, duration)(csr)
+	}
+}
+
+func extractCAdata(caCertData, caKeyData []byte) ([]byte, []byte, error) {
+	certBlock, _ := pem.Decode(caCertData)
+	caCert, err := x509.ParseCertificate(certBlock.Bytes)
+	if err != nil {
+		return nil, nil, errors.Wrapf(err, "failed to parse ca certificate")
+	}
+	keyBlock, _ := pem.Decode(caKeyData)
+	caKey, err := x509.ParsePKCS8PrivateKey(keyBlock.Bytes)
+	if err != nil {
+		return nil, nil, errors.Wrapf(err, "failed to parse ca key")
+	}
+
+	caConfig := &openshiftcrypto.TLSCertificateConfig{
+		Certs: []*x509.Certificate{caCert},
+		Key:   caKey,
+	}
+	return caConfig.GetPEMBytes()
+}
+
+// TemplatePermissionConfigFunc returns a func that can grant permission for addon agent
+// that is deployed by addon template.
+// the returned func will create a rolebinding to bind the clusterRole/role which is
+// specified by the user, so the user is required to make sure the existence of the
+// clusterRole/role
+func TemplatePermissionConfigFunc(
+	addonName string,
+	templateGetter AddonTemplateGetterFunc,
+	hubKubeClient kubernetes.Interface,
+	rolebindingLister rbacv1lister.RoleBindingLister,
+) agent.PermissionConfigFunc {
+
+	return func(cluster *clusterv1.ManagedCluster, addon *addonapiv1alpha1.ManagedClusterAddOn) error {
+		template, err := templateGetter(addon, cluster.Name, addonName)
+		if err != nil {
+			return err
+		}
+		if template == nil {
+			return nil
+		}
+
+		for _, registration := range template.Spec.Registration {
+			switch registration.Type {
+			case addonapiv1alpha1.RegistrationTypeKubeClient:
+				kcrc := registration.KubeClient
+				if kcrc == nil {
+					continue
+				}
+
+				err := createKubeClientPermissions(kcrc, cluster, addon, hubKubeClient, rolebindingLister)
+				if err != nil {
+					return err
+				}
+
+			case addonapiv1alpha1.RegistrationTypeCustomSigner:
+				continue
+
+			default:
+				utilruntime.HandleError(fmt.Errorf("unsupported registration type %s", registration.Type))
+			}
+
+		}
+
+		return nil
+	}
+}
+
+func createKubeClientPermissions(
+	kcrc *addonapiv1alpha1.KubeClientRegistrationConfig,
+	cluster *clusterv1.ManagedCluster,
+	addon *addonapiv1alpha1.ManagedClusterAddOn,
+	hubKubeClient kubernetes.Interface,
+	rolebindingLister rbacv1lister.RoleBindingLister) error {
+
+	for _, pc := range kcrc.HubPermissions {
+		switch pc.Type {
+		case addonapiv1alpha1.HubPermissionsBindingCurrentCluster:
+			klog.V(5).Infof("Set hub permission for addon %s/%s, UID: %s, APIVersion: %s, Kind: %s",
+				addon.Namespace, addon.Name, addon.UID, addon.APIVersion, addon.Kind)
+
+			owner := metav1.OwnerReference{
+				// TODO: use apiVersion and kind in addon object, but now they could be empty at some unknown reason
+				APIVersion: "addon.open-cluster-management.io/v1alpha1",
+				Kind:       "ManagedClusterAddOn",
+				Name:       addon.Name,
+				UID:        addon.UID,
+			}
+			err := createPermissionBinding(hubKubeClient, rolebindingLister,
+				cluster.Name, addon.Name, cluster.Name, pc.RoleRef, owner)
+			if err != nil {
+				return err
+			}
+		case addonapiv1alpha1.HubPermissionsBindingSingleNamespace:
+			if pc.SingleNamespace == nil {
+				return fmt.Errorf("single namespace is nil")
+			}
+			owner := metav1.OwnerReference{
+				APIVersion: cluster.APIVersion,
+				Kind:       cluster.Kind,
+				Name:       cluster.Name,
+				UID:        cluster.UID,
+			}
+			err := createPermissionBinding(hubKubeClient, rolebindingLister,
+				cluster.Name, addon.Name, pc.SingleNamespace.Namespace, pc.RoleRef, owner)
+			if err != nil {
+				return err
+			}
+		}
+	}
+	return nil
+}
+
+func createPermissionBinding(hubKubeClient kubernetes.Interface, rolebindingLister rbacv1lister.RoleBindingLister,
+	clusterName, addonName, namespace string, roleRef rbacv1.RoleRef, owner metav1.OwnerReference) error {
+	// TODO: confirm the group
+	groups := agent.DefaultGroups(clusterName, addonName)
+	subject := []rbacv1.Subject{}
+	for _, group := range groups {
+		subject = append(subject, rbacv1.Subject{
+			Kind: "Group", APIGroup: "rbac.authorization.k8s.io", Name: group,
+		})
+	}
+	binding := &rbacv1.RoleBinding{
+		ObjectMeta: metav1.ObjectMeta{
+			Name: fmt.Sprintf("open-cluster-management:%s:%s:agent",
+				addonName, strings.ToLower(roleRef.Kind)),
+			Namespace:       namespace,
+			OwnerReferences: []metav1.OwnerReference{owner},
+			Labels: map[string]string{
+				addonapiv1alpha1.AddonLabelKey: addonName,
+			},
+		},
+		RoleRef:  roleRef,
+		Subjects: subject,
+	}
+
+	_, err := rolebindingLister.RoleBindings(namespace).Get(binding.Name)
+	switch {
+	case apierrors.IsNotFound(err):
+		_, createErr := hubKubeClient.RbacV1().RoleBindings(namespace).Create(context.TODO(), binding, metav1.CreateOptions{})
+		if createErr != nil && !apierrors.IsAlreadyExists(createErr) {
+			return createErr
+		}
+	case err != nil:
+		return err
+	default:
+		klog.Infof("rolebinding %s already exists", binding.Name)
+	}
+
+	// TODO: update the rolebinding if it is not the same
+	return nil
+}
