+++ pkg/manager/templateagent/testmanifests/addontemplate.yaml
