+++ pkg/utils/csr_helpers.go
@@ -1,7 +1,6 @@
 package utils
 
 import (
-	"context"
 	"crypto/rand"
 	"crypto/rsa"
 	"crypto/x509"
@@ -11,27 +10,17 @@ import (
 	"strings"
 	"time"
 
-	openshiftcrypto "github.com/openshift/library-go/pkg/crypto"
-	"github.com/pkg/errors"
 	certificatesv1 "k8s.io/api/certificates/v1"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime/schema"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/apimachinery/pkg/util/sets"
 	"k8s.io/client-go/discovery/cached/memory"
 	"k8s.io/client-go/kubernetes"
 	"k8s.io/client-go/restmapper"
 	"k8s.io/klog/v2"
 	addonapiv1alpha1 "open-cluster-management.io/api/addon/v1alpha1"
-	addonlisterv1alpha1 "open-cluster-management.io/api/client/addon/listers/addon/v1alpha1"
 	clusterv1 "open-cluster-management.io/api/cluster/v1"
 
 	"open-cluster-management.io/addon-framework/pkg/agent"
-)
-
-const (
-	TLSCACert = "ca.crt"
-	TLSCAKey  = "ca.key"
 )
 
 var serialNumberLimit = new(big.Int).Lsh(big.NewInt(1), 128)
@@ -63,56 +52,6 @@ func DefaultSignerWithExpiry(caKey, caData []byte, duration time.Duration) agent
 		}
 		return data
 	}
-}
-
-func CustomSignerWithExpiry(
-	kubeclient kubernetes.Interface,
-	customSignerConfig *addonapiv1alpha1.CustomSignerRegistrationConfig,
-	duration time.Duration) agent.CSRSignerFunc {
-	return func(csr *certificatesv1.CertificateSigningRequest) []byte {
-		if customSignerConfig == nil {
-			utilruntime.HandleError(fmt.Errorf("custome signer is nil"))
-			return nil
-		}
-
-		if csr.Spec.SignerName != customSignerConfig.SignerName {
-			return nil
-		}
-		caSecret, err := kubeclient.CoreV1().Secrets(customSignerConfig.SigningCA.Namespace).Get(
-			context.TODO(), customSignerConfig.SigningCA.Name, metav1.GetOptions{})
-		if err != nil {
-			utilruntime.HandleError(fmt.Errorf("get custome signer ca %s/%s failed, %v",
-				customSignerConfig.SigningCA.Namespace, customSignerConfig.SigningCA.Name, err))
-			return nil
-		}
-
-		caData, caKey, err := extractCAdata(caSecret.Data[TLSCACert], caSecret.Data[TLSCAKey])
-		if customSignerConfig == nil {
-			utilruntime.HandleError(fmt.Errorf("get ca %s/%s data failed, %v",
-				customSignerConfig.SigningCA.Namespace, customSignerConfig.SigningCA.Name, err))
-			return nil
-		}
-		return DefaultSignerWithExpiry(caKey, caData, duration)(csr)
-	}
-}
-
-func extractCAdata(caCertData, caKeyData []byte) ([]byte, []byte, error) {
-	certBlock, _ := pem.Decode(caCertData)
-	caCert, err := x509.ParseCertificate(certBlock.Bytes)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to parse ca certificate")
-	}
-	keyBlock, _ := pem.Decode(caKeyData)
-	caKey, err := x509.ParsePKCS8PrivateKey(keyBlock.Bytes)
-	if err != nil {
-		return nil, nil, errors.Wrapf(err, "failed to parse ca key")
-	}
-
-	caConfig := &openshiftcrypto.TLSCertificateConfig{
-		Certs: []*x509.Certificate{caCert},
-		Key:   caKey,
-	}
-	return caConfig.GetPEMBytes()
 }
 
 func signCSR(csr *certificatesv1.CertificateSigningRequest, caCert *x509.Certificate, caKey *rsa.PrivateKey, duration time.Duration) ([]byte, error) {
@@ -181,20 +120,6 @@ func parseCSR(pemBytes []byte) (*x509.CertificateRequest, error) {
 	return csr, nil
 }
 
-// KubeClientCSRApprover approve the csr when addon agent uses default group, default user and
-// "kubernetes.io/kube-apiserver-client" signer to sign csr.
-func KubeClientCSRApprover(agentName string) agent.CSRApproveFunc {
-	return func(
-		cluster *clusterv1.ManagedCluster,
-		addon *addonapiv1alpha1.ManagedClusterAddOn,
-		csr *certificatesv1.CertificateSigningRequest) bool {
-		if csr.Spec.SignerName != certificatesv1.KubeAPIServerClientSignerName {
-			return false
-		}
-		return DefaultCSRApprover(agentName)(cluster, addon, csr)
-	}
-}
-
 // DefaultCSRApprover approve the csr when addon agent uses default group and default user to sign csr.
 func DefaultCSRApprover(agentName string) agent.CSRApproveFunc {
 	return func(
@@ -247,15 +172,15 @@ func DefaultCSRApprover(agentName string) agent.CSRApproveFunc {
 	}
 }
 
-// CustomerSignerCSRApprover approve the csr when addon agent uses custom signer to sign csr.
-func CustomerSignerCSRApprover(agentName string) agent.CSRApproveFunc {
-	return func(
-		cluster *clusterv1.ManagedCluster,
-		addon *addonapiv1alpha1.ManagedClusterAddOn,
-		csr *certificatesv1.CertificateSigningRequest) bool {
+// UnionCSRApprover is a union func for multiple approvers
+func UnionCSRApprover(approvers ...agent.CSRApproveFunc) agent.CSRApproveFunc {
+	return func(cluster *clusterv1.ManagedCluster, addon *addonapiv1alpha1.ManagedClusterAddOn, csr *certificatesv1.CertificateSigningRequest) bool {
+		for _, approver := range approvers {
+			if !approver(cluster, addon, csr) {
+				return false
+			}
+		}
 
-		klog.Infof("Customer signer CSR is approved. cluster: %s, addon %s, requester: %s",
-			cluster.Name, addon.Name, csr.Spec.Username)
 		return true
 	}
 }
@@ -283,195 +208,4 @@ func IsCSRSupported(nativeClient kubernetes.Interface) (bool, bool, error) {
 		}
 	}
 	return v1CSRSupported, v1beta1CSRSupported, nil
-}
-
-// AddonTemplateGetterFunc returns the desired addon template for the addon, if addon is nil, the clusterName and
-// addonName should not be empty
-type AddonTemplateGetterFunc func(addon *addonapiv1alpha1.ManagedClusterAddOn, clusterName, addonName string) (*addonapiv1alpha1.AddOnTemplate, error)
-
-// DefaultDesiredAddonTemplateGetter returns a function that returns the desired addon template
-// if addon is nil, it will get the addon from cache, then get the desired template by addon status
-func DefaultDesiredAddonTemplateGetter(
-	mcaLister addonlisterv1alpha1.ManagedClusterAddOnLister,
-	atLister addonlisterv1alpha1.AddOnTemplateLister) AddonTemplateGetterFunc {
-	return func(addon *addonapiv1alpha1.ManagedClusterAddOn, clusterName, addonName string) (*addonapiv1alpha1.AddOnTemplate, error) {
-
-		if addon == nil {
-			var err error
-			addon, err = mcaLister.ManagedClusterAddOns(clusterName).Get(addonName)
-			if err != nil {
-				return nil, err
-			}
-		}
-
-		return GetDesiredAddOnTemplate(atLister, addon)
-	}
-}
-
-func TemplateCSRConfigurationsFunc(
-	addonName, agentName string,
-	templateGetter AddonTemplateGetterFunc,
-) func(cluster *clusterv1.ManagedCluster) []addonapiv1alpha1.RegistrationConfig {
-
-	return func(cluster *clusterv1.ManagedCluster) []addonapiv1alpha1.RegistrationConfig {
-		template, err := templateGetter(nil, cluster.Name, addonName)
-		if err != nil {
-			utilruntime.HandleError(fmt.Errorf("failed to get addon %s template: %v", addonName, err))
-			return nil
-		}
-		if template == nil {
-			return nil
-		}
-
-		contain := func(rcs []addonapiv1alpha1.RegistrationConfig, signerName string) bool {
-			for _, rc := range rcs {
-				if rc.SignerName == signerName {
-					return true
-				}
-			}
-			return false
-		}
-
-		registrationConfigs := make([]addonapiv1alpha1.RegistrationConfig, 0)
-		for _, registration := range template.Spec.Registration {
-			switch registration.Type {
-			case addonapiv1alpha1.RegistrationTypeKubeClient:
-				if !contain(registrationConfigs, certificatesv1.KubeAPIServerClientSignerName) {
-					configs := agent.KubeClientSignerConfigurations(addonName, agentName)(cluster)
-					registrationConfigs = append(registrationConfigs, configs...)
-				}
-
-			case addonapiv1alpha1.RegistrationTypeCustomSigner:
-				if registration.CustomSigner == nil {
-					continue
-				}
-				if !contain(registrationConfigs, registration.CustomSigner.SignerName) {
-					configs := agent.CustomSignerConfigurations(
-						addonName, agentName, registration.CustomSigner)(cluster)
-					registrationConfigs = append(registrationConfigs, configs...)
-				}
-
-			default:
-				utilruntime.HandleError(fmt.Errorf("unsupported registration type %s", registration.Type))
-			}
-
-		}
-
-		return registrationConfigs
-	}
-}
-
-func TemplateCSRApproveCheckFunc(
-	addonName, agentName string,
-	templateGetter AddonTemplateGetterFunc,
-) agent.CSRApproveFunc {
-
-	return func(cluster *clusterv1.ManagedCluster, addon *addonapiv1alpha1.ManagedClusterAddOn,
-		csr *certificatesv1.CertificateSigningRequest) bool {
-
-		template, err := templateGetter(addon, cluster.Name, addonName)
-		if err != nil {
-			utilruntime.HandleError(fmt.Errorf("failed to get addon %s template: %v", addonName, err))
-			return false
-		}
-		if template == nil {
-			return false
-		}
-
-		for _, registration := range template.Spec.Registration {
-			switch registration.Type {
-			case addonapiv1alpha1.RegistrationTypeKubeClient:
-
-				if csr.Spec.SignerName == certificatesv1.KubeAPIServerClientSignerName {
-					return KubeClientCSRApprover(agentName)(cluster, addon, csr)
-				}
-
-			case addonapiv1alpha1.RegistrationTypeCustomSigner:
-				if registration.CustomSigner == nil {
-					continue
-				}
-				if csr.Spec.SignerName == registration.CustomSigner.SignerName {
-					return CustomerSignerCSRApprover(addonName)(cluster, addon, csr)
-				}
-
-			default:
-				utilruntime.HandleError(fmt.Errorf("unsupported registration type %s", registration.Type))
-			}
-
-		}
-
-		return false
-	}
-}
-
-func TemplateCSRSignFunc(
-	addonName, agentName string,
-	templateGetter AddonTemplateGetterFunc,
-	hubKubeClient kubernetes.Interface,
-) agent.CSRSignerFunc {
-
-	return func(csr *certificatesv1.CertificateSigningRequest) []byte {
-		// TODO: consider to change the agent.CSRSignerFun to accept parameter addon
-		getClusterName := func(userName string) string {
-			// the common name of csr is in format of "system:open-cluster-management:{clusterName}:{id}"
-			// get the cluster name from the common name
-			trimedCommonName := strings.TrimPrefix(userName, "system:open-cluster-management:")
-			return strings.Split(trimedCommonName, ":")[0]
-		}
-
-		clusterName := getClusterName(csr.Spec.Username)
-		template, err := templateGetter(nil, clusterName, addonName)
-		if err != nil {
-			utilruntime.HandleError(fmt.Errorf("failed to get addon %s template: %v", addonName, err))
-			return nil
-		}
-		if template == nil {
-			return nil
-		}
-
-		for _, registration := range template.Spec.Registration {
-			switch registration.Type {
-			case addonapiv1alpha1.RegistrationTypeKubeClient:
-				continue
-
-			case addonapiv1alpha1.RegistrationTypeCustomSigner:
-				if registration.CustomSigner == nil {
-					continue
-				}
-				if csr.Spec.SignerName == registration.CustomSigner.SignerName {
-					return CustomSignerWithExpiry(hubKubeClient, registration.CustomSigner, 24*time.Hour)(csr)
-				}
-
-			default:
-				utilruntime.HandleError(fmt.Errorf("unsupported registration type %s", registration.Type))
-			}
-
-		}
-
-		return nil
-	}
-}
-
-// GetDesiredAddOnTemplate returns the desired template of the addon
-func GetDesiredAddOnTemplate(
-	atLister addonlisterv1alpha1.AddOnTemplateLister,
-	addon *addonapiv1alpha1.ManagedClusterAddOn) (*addonapiv1alpha1.AddOnTemplate, error) {
-	ok, templateRef := AddonTemplateConfigRef(addon.Status.ConfigReferences)
-	if !ok {
-		klog.V(4).Infof("Addon %s template config in status is empty", addon.Name)
-		return nil, nil
-	}
-
-	desiredTemplate := templateRef.DesiredConfig
-	if desiredTemplate == nil || desiredTemplate.SpecHash == "" {
-		klog.Infof("Addon %s template spec hash is empty", addon.Name)
-		return nil, fmt.Errorf("addon %s template desired spec hash is empty", addon.Name)
-	}
-
-	template, err := atLister.Get(desiredTemplate.Name)
-	if err != nil {
-		return nil, err
-	}
-
-	return template.DeepCopy(), nil
 }
