+++ pkg/utils/helpers.go
@@ -3,6 +3,7 @@ package utils
 import (
 	"bytes"
 	"context"
+	"crypto/sha256"
 	"encoding/json"
 	"fmt"
 	"reflect"
@@ -14,6 +15,7 @@ import (
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	"k8s.io/apimachinery/pkg/api/meta"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
+	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 	"k8s.io/apimachinery/pkg/types"
 	coreclientv1 "k8s.io/client-go/kubernetes/typed/core/v1"
@@ -346,4 +348,24 @@ func IsOwnedByCMA(addon *addonapiv1alpha1.ManagedClusterAddOn) bool {
 		return true
 	}
 	return false
+}
+
+// GetSpecHash returns the sha256 hash of the spec field of the given object
+func GetSpecHash(obj *unstructured.Unstructured) (string, error) {
+	if obj == nil {
+		return "", fmt.Errorf("object is nil")
+	}
+	spec, ok := obj.Object["spec"]
+	if !ok {
+		return "", fmt.Errorf("object has no spec field")
+	}
+
+	specBytes, err := json.Marshal(spec)
+	if err != nil {
+		return "", err
+	}
+
+	hash := sha256.Sum256(specBytes)
+
+	return fmt.Sprintf("%x", hash), nil
 }
