+++ pkg/utils/permission.go
@@ -3,16 +3,13 @@ package utils
 import (
 	"context"
 	"fmt"
-	"strings"
 
 	rbacv1 "k8s.io/api/rbac/v1"
 	"k8s.io/apimachinery/pkg/api/equality"
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	utilruntime "k8s.io/apimachinery/pkg/util/runtime"
 	"k8s.io/client-go/kubernetes"
 	rbacclientv1 "k8s.io/client-go/kubernetes/typed/rbac/v1"
-	"k8s.io/klog/v2"
 	"k8s.io/utils/pointer"
 	addonapiv1alpha1 "open-cluster-management.io/api/addon/v1alpha1"
 	clusterv1 "open-cluster-management.io/api/cluster/v1"
@@ -354,118 +351,4 @@ func ApplyRoleBinding(ctx context.Context, client rbacclientv1.RoleBindingsGette
 
 	actual, err := client.RoleBindings(requiredCopy.Namespace).Update(ctx, existingCopy, metav1.UpdateOptions{})
 	return actual, true, err
-}
-
-// TemplatePermissionConfigFunc returns a func that can grant permission for addon agent
-// that is deployed by addon template.
-// the returned func will create a rolebinding to bind the clusterRole/role which is
-// specified by the user, so the user is required to make sure the existence of the
-// clusterRole/role
-func TemplatePermissionConfigFunc(
-	addonName string,
-	templateGetter AddonTemplateGetterFunc,
-	hubKubeClient kubernetes.Interface,
-) agent.PermissionConfigFunc {
-
-	return func(cluster *clusterv1.ManagedCluster, addon *addonapiv1alpha1.ManagedClusterAddOn) error {
-		template, err := templateGetter(addon, cluster.Name, addonName)
-		if err != nil {
-			return err
-		}
-		if template == nil {
-			return nil
-		}
-
-		for _, registration := range template.Spec.Registration {
-			switch registration.Type {
-			case addonapiv1alpha1.RegistrationTypeKubeClient:
-
-				kcrc := registration.KubeClient
-				if kcrc == nil {
-					continue
-				}
-
-				for _, pc := range kcrc.HubPermissions {
-					switch pc.Type {
-					case addonapiv1alpha1.HubPermissionsBindingCurrentCluster:
-						klog.V(5).Infof("Set hub permission for addon %s/%s, UID: %s, APIVersion: %s, Kind: %s",
-							addon.Namespace, addon.Name, addon.UID, addon.APIVersion, addon.Kind)
-
-						owner := metav1.OwnerReference{
-							// TODO: use apiVersion and kind in addon object, but now they could be empty at some unknown reason
-							APIVersion: "addon.open-cluster-management.io/v1alpha1",
-							Kind:       "ManagedClusterAddOn",
-							Name:       addon.Name,
-							UID:        addon.UID,
-						}
-						err := createPermissionBinding(hubKubeClient,
-							cluster.Name, addon.Name, cluster.Name, pc.RoleRef, owner)
-						if err != nil {
-							return err
-						}
-					case addonapiv1alpha1.HubPermissionsBindingSingleNamespace:
-						if pc.SingleNamespace == nil {
-							return fmt.Errorf("single namespace is nil")
-						}
-						owner := metav1.OwnerReference{
-							APIVersion: cluster.APIVersion,
-							Kind:       cluster.Kind,
-							Name:       cluster.Name,
-							UID:        cluster.UID,
-						}
-						err := createPermissionBinding(hubKubeClient,
-							cluster.Name, addon.Name, pc.SingleNamespace.Namespace, pc.RoleRef, owner)
-						if err != nil {
-							return err
-						}
-					}
-				}
-
-			case addonapiv1alpha1.RegistrationTypeCustomSigner:
-				continue
-
-			default:
-				utilruntime.HandleError(fmt.Errorf("unsupported registration type %s", registration.Type))
-			}
-
-		}
-
-		return nil
-	}
-}
-
-func createPermissionBinding(hubKubeClient kubernetes.Interface,
-	clusterName, addonName, namespace string, roleRef rbacv1.RoleRef, owner metav1.OwnerReference) error {
-	// TODO: confirm the group
-	groups := agent.DefaultGroups(clusterName, addonName)
-	subject := []rbacv1.Subject{}
-	for _, group := range groups {
-		subject = append(subject, rbacv1.Subject{
-			Kind: "Group", APIGroup: "rbac.authorization.k8s.io", Name: group,
-		})
-	}
-	binding := &rbacv1.RoleBinding{
-		ObjectMeta: metav1.ObjectMeta{
-			Name: fmt.Sprintf("open-cluster-management:%s:%s:agent",
-				addonName, strings.ToLower(roleRef.Kind)),
-			Namespace:       namespace,
-			OwnerReferences: []metav1.OwnerReference{owner},
-		},
-		RoleRef:  roleRef,
-		Subjects: subject,
-	}
-
-	_, err := hubKubeClient.RbacV1().RoleBindings(namespace).Get(context.TODO(), binding.Name, metav1.GetOptions{})
-	switch {
-	case apierrors.IsNotFound(err):
-		_, createErr := hubKubeClient.RbacV1().RoleBindings(namespace).Create(context.TODO(), binding, metav1.CreateOptions{})
-		if createErr != nil && !apierrors.IsAlreadyExists(createErr) {
-			return createErr
-		}
-	case err != nil:
-		return err
-	}
-
-	// TODO: update the rolebinding if it is not the same
-	return nil
 }
